Add webpack and split modules into separate files
