--- assets/js/main.js
@@ -1,196 +1,0 @@
-const createPlayer = (name, shape) => ({
-  getName() {
-    return name;
-  },
-  getShape() {
-    return shape;
-  },
-});
-
-const display = (() => {
-  const $form = document.querySelector('#player');
-
-  const showPlayerForm = (heading) => {
-    $form.querySelector('h1').innerHTML = heading;
-    $form.classList.add('display-block');
-  };
-
-  const hidePlayerForm = () => {
-    $form.classList.remove('display-block');
-    $form.classList.add('display-none');
-  };
-
-  const getPlayerFormData = () => new Promise((resolve) => {
-    function onSubmit(e) {
-      e.preventDefault();
-      const $name = $form.querySelector('#player_name');
-      resolve({
-        name: $name.value,
-      });
-      $name.value = '';
-      $form.removeEventListener('submit', onSubmit);
-    }
-
-    $form.addEventListener('submit', onSubmit);
-  });
-
-  return {
-    showPlayerForm,
-    getPlayerFormData,
-    hidePlayerForm,
-  };
-})();
-
-const gameboard = (() => {
-  let gameCells = ['', '', '', '', '', '', '', '', ''];
-  const $boardDiv = document.querySelector('.board-container');
-  const $boardCells = $boardDiv.querySelectorAll('.cell');
-  const $gameStateDiv = document.querySelector('#game-state');
-
-  let clickCallback;
-
-  const fillPlayedCell = (index, shape) => {
-    if (gameCells[index] === '') {
-      gameCells[index] = shape;
-      $boardDiv.querySelector(`#cell-${index}`).innerHTML = shape;
-      return true;
-    }
-    return false;
-  };
-
-  const checkWinner = (shape) => {
-    const winningCells = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6],
-    ];
-
-    const playerBoard = [];
-
-    for (let i = 0; i < gameCells.length; i += 1) {
-      if (gameCells[i] === shape) {
-        playerBoard.push(i);
-      }
-    }
-
-    for (let i = 0; i < winningCells.length; i += 1) {
-      const cells = gameCells;
-      if (winningCells[i].every(index => cells[index] === shape)) return 'win';
-    }
-
-    for (let i = 0; i < gameCells.length; i += 1) {
-      if (gameCells[i] === '') {
-        return 'nowin';
-      }
-    }
-
-    return 'draw';
-  };
-
-  const showBoard = () => {
-    $boardDiv.classList.add('display-block');
-  };
-
-  const updateGameStateDisplay = (str) => {
-    $gameStateDiv.innerHTML = str;
-  };
-
-  const onCellClick = (cb) => {
-    clickCallback = cb;
-    $boardCells.forEach((cell) => {
-      cell.addEventListener('click', cb);
-    });
-  };
-
-  const clearClickListener = () => {
-    $boardCells.forEach((cell) => {
-      cell.removeEventListener('click', clickCallback);
-    });
-  };
-
-  const clear = () => {
-    $boardCells.forEach((div) => {
-      div.innerHTML = '';
-    });
-    gameCells = ['', '', '', '', '', '', '', '', ''];
-  };
-
-  return {
-    showBoard,
-    onCellClick,
-    fillPlayedCell,
-    checkWinner,
-    updateGameStateDisplay,
-    clearClickListener,
-    clear,
-  };
-})();
-
-const game = (() => {
-  const $playAgain = document.querySelector('#play-again');
-  let player1 = createPlayer('Test 1', 'X');
-  let player2 = createPlayer('Test 2', 'O');
-  let currentPlayer = player1;
-
-  const showPlayAgain = () => {
-    $playAgain.classList.add('display-block');
-  };
-
-  function handleTurn(cell) {
-    if (!gameboard.fillPlayedCell(cell, currentPlayer.getShape())) return;
-
-    switch (gameboard.checkWinner(currentPlayer.getShape())) {
-      case 'win':
-        gameboard.updateGameStateDisplay(`${currentPlayer.getName()} has won.`);
-        showPlayAgain();
-        gameboard.clearClickListener();
-        return;
-      case 'draw':
-        gameboard.updateGameStateDisplay('It\'s a draw.');
-        showPlayAgain();
-        gameboard.clearClickListener();
-        return;
-      default:
-        break;
-    }
-
-    currentPlayer = currentPlayer === player1 ? player2 : player1;
-    gameboard.updateGameStateDisplay(`${currentPlayer.getName()}'s turn.`);
-  }
-
-  async function init() {
-    let data = null;
-    display.showPlayerForm('Player 1 details');
-    data = await display.getPlayerFormData();
-    player1 = createPlayer(data.name, 'X');
-    display.showPlayerForm('Player 2 details');
-    data = await display.getPlayerFormData();
-    player2 = createPlayer(data.name, 'O');
-    display.hidePlayerForm();
-    currentPlayer = player1;
-
-    gameboard.showBoard();
-    gameboard.onCellClick((e) => handleTurn(+e.target.dataset.cell));
-    gameboard.updateGameStateDisplay(`${currentPlayer.getName()}'s turn.`);
-  }
-
-  $playAgain.addEventListener('click', () => {
-    gameboard.clear();
-    gameboard.onCellClick((e) => handleTurn(+e.target.dataset.cell));
-    gameboard.updateGameStateDisplay(`${currentPlayer.getName()}'s turn.`);
-  });
-
-  return {
-    init,
-  };
-})();
-
-game.init().catch((err) => {
-  // eslint-disable-next-line no-console
-  console.error(err);
-});
