+++ index.html
@@ -31,6 +31,6 @@
         <button id="play-again">Play again?</button>
     </div>
 
-    <script src="assets/js/main.js"></script>
+    <script src="assets/js/bundle.min.js"></script>
 </body>
 </html>
