+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "lint": "eslint .",
     "lint:fix": "eslint --fix .",
-    "lint:style": "stylelint **/*.{css,scss}"
+    "lint:style": "stylelint **/*.{css,scss}",
+    "build": "webpack --mode production",
+    "start": "webpack serve"
   },
   "repository": {
     "type": "git",
@@ -26,6 +28,9 @@
     "stylelint": "^13.7.2",
     "stylelint-config-standard": "^20.0.0",
     "stylelint-csstree-validator": "^1.8.0",
-    "stylelint-scss": "^3.18.0"
+    "stylelint-scss": "^3.18.0",
+    "webpack": "^5.1.3",
+    "webpack-cli": "^4.1.0",
+    "webpack-dev-server": "^3.11.0"
   }
 }
