+++ src/player.js
@@ -1,0 +1,10 @@
+const createPlayer = (name, shape) => ({
+  getName() {
+    return name;
+  },
+  getShape() {
+    return shape;
+  },
+});
+
+export default createPlayer;
